Returning JsonSchema from the database and creating a new SQLSchema row.

+++ src/main/java/com/parola/document2sql/mapper/repository/JsonSchemaRepository.java
@@ -3,5 +3,5 @@ package com.parola.document2sql.mapper.repository;
 import com.parola.document2sql.mapper.entity.JsonSchema;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface JsonSchemaRepository extends ListCrudRepository<JsonSchema,Integer> {
+public interface JsonSchemaRepository extends ListCrudRepository<JsonSchema,Long> {
 }
